Expose absolute file of project under test
Absolute file provides a better API to access any projects file and does
not require the user to construct files itself.

`GradleProject.buildDirectory()` was removed because it exposed a single
directory and essentially meant that it has to be called repeatedly in
order to access inner files. Gradle directory structure is well defined
thus it's ubiquitous to access files directly by path.